backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==2.0.12
cryptography==43.0.3
defusedxml==0.8.0rc2
Django==3.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
django-cors-headers==3.13.0
djoser==2.3.1
idna==3.10
oauthlib==3.2.2
Pillow==9.0.0
pycparser==2.22
PyJWT==2.10.0
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2024.2
requests==2.26.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==1.26.20
psycopg2-binary==2.9.3
django-filter==23.1
isort==5.13.2
tqdm==4.67.1